LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/the-video-element/intrinsic_sizes-expected.txt

spec reference

    

PASS default object size is 300x150 
PASS default height is half the width 
PASS default width is twice the height 
FAIL default object size after src is removed assert_equals: expected "300px" but got "320px"
FAIL default object size after poster is removed assert_equals: expected "102px" but got "300px"
